backend/requirements.txt
# Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database & ORM
sqlalchemy>=2.0.0
aiomysql>=0.2.0
alembic>=1.11.0
mysqlclient>=2.2.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Security (Auth)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt==4.0.1

# Redis & Rate Limiting (수정됨)
redis>=5.0.0
fastapi-limiter>=0.1.5

# Test
pytest>=7.4.0
httpx>=0.24.1
pytest-asyncio>=0.21.0
aiosqlite>=0.19.0

firebase-admin>=6.2.0

Faker>=19.0.0

python-multipart